Fix crash in-flight if hj-planner has never opened
When the amount of fuel changes on the ship, the function of rebuilding
the route strings is called, because colors can change there. The
function uses a pointer to the player, which is calculated when the
planner window is displayed. If the window never been shown, a reference
to the null pointer occurs and the game crashes.