Another 2 fixes to the sector map
- fix that Auto Route button crashes game when no hyperdrive fitted. Add
3 possible options - successful route building, no drive, no valid
route. Also turn the last static function in LuaSectorView.cpp into a
lambda, because it turned out that it was not needed in other functions.

- fix that when moving the SectorView camera by entering coordinates to
the search field, the camera position gets "locked" to the coordinates
and you can't move it (now it moves only when the search bar changes)