Refactor IniConfig, Input optionally saves config
IniConfig::Save() writes the config to the file it was Read() from.
GameConfig is reduced to defaults only
Several Input methods now maintain config state as well as input class state

Set default screen resolution to 1280x720

This is an opinionated default, but given that >90% of our userbase has
a widescreen monitor (and we're getting rid of the 800x600 unit metric)
it's time to move on.

Fix missing config file preventing saving config