Expose InputFrame 2.0 to Lua
Querying bindings is now exposed to Lua, with the added side effect of
slightly simplifying our binding menu control flow. I've refactored the
existing binding window to support the new API, although it's a faint
shadow of what's possible under the new system.

Further work (in another PR) will be required to support binding of
multiple key chords from the GUI; this may well be lumped in with the
rewrite of the Lua console.