It is the nature of all things to die...
and the reaper's due must be paid.

As we fly back to base after another successful mission, let us clear
comms in a moment of silence for the wingmates who no longer fly in
this squadron.

- To Robn, who wrote newUI in the first place. Without you, there would
  be no newUI or pigui, and I'd still be complaining about oldUI.
- To Johnbartholomew, who spearheaded the move of existing systems to
  newUI and implemented plenty of code we still use today.
- To ecraven, who did all of the work of adding ImGui to the code base
 and worked tirelessly to port existing screens.
- To vakhoir, who not only worked tirelessly to port the most
  recalcitrant of screens, but developed UI frameworks and paradigms
  that made working on the UI a pleasure instead of a pain.

This commit totally removes the newUI from the codebase. Everything that
made reference to it (save the LuaConsole) has been ported or struck
from the code. This has simplified some systems significantly, has
allowed us to drop 8,000+ lines of C++ code, cut our build times
significantly, and brings us one step closer to a modern rendering and
tools pipeline.