CMake: Fix modelcompiler not compiling anything
When models are installed on the system, the modelcompiler would use
these instead of the local ones found in the "data" folder. However,
the *.model files are not installed on the system, so the
modelcompiler would end up not compiling anything.

Fix this by returning only the local "data" folder files in
FileSystemPosix.cpp if the PIONEER_LOCAL_DATA_ONLY environment variable
is set, and set this environment variable in CMake when the
modelcompiler is called.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>