Reduce probability for imported goods to be sold out
In general: Just because something is imported, doesn't mean it will be sold
out, rather that depends on demand exceeding supply (which can happen for
imported and exported products).

For the likelihood of outcome computed from the probability distribution,
consider: although it is uniform, shifting it one unit left/right gives
no-uniform change in likelihood, since the distribution not uniform.