Move the surface starport system bodies on loading
Sometimes surface ports move when terrain is generated because it is not
suitable for their original position.  When serialized, the physical
body saves its changed position, but the system body does not.
Therefore, upon subsequent loading, the position of the system body and
the physical body may not coincide. In this commit, the same mechanism
is launched as during the initial generation of the relief, which
displaces the system bodies to the right position.