Reduce the flow of tradeships
- reduce the flow to the most popular station in system by 40%.
- reduce the maximum station load by 25%
- reduce the maximum ships limit by 40%

This is to improve the performance. although these parameters are
independent, it is better to change them synchronously.  For example, if
you just reduce the allowable number of ships, in all systems this
number can become the same.  If the load on the station is not reduced,
the ships will stay there for too long.