Fix NewsEvent not doing it's thing after save/load cycle
The cargo to have its stock and price changed was saved as a EqupType object,
which is not a "deep save", but will have a new memory hash every new game,
thus price and stock for a commodity that was subject to high/low demand, was
unchanged.

This should be backwards compatible with older saves.