Render PiGui using Pioneer's Renderer API
Instead of generating a bunch of raw OpenGL commands, use the pioneer renderer.

This allows us to avoid setting and resetting a bunch of OpenGL state per frame, and we can be more efficient with buffer transfers as well.

This fixed a longstanding bug where we would leak previous versions of font textures.