LuaObject: document an hypothetical ApplyAt method
This is an RFC API aiming to mitigate the safety problems caused by
Timer:CallAt as seen in plenty of issues, the latest of which is #5220.
The fundamental issue with the CallAt API is that the environment in
which the closure is called can be drastically different to the one in
which it was defined.

This new API aims to guarantee the caller that at least one object must
still be valid when calling the closure, by storing it out of band and
having the engine check its validity before calling the callback.

This would allow some simplification in existing correct callbacks, e.g.
in the assassination module, we could turn this

    Timer:CallAt(mission.due, function () if mission.ship:exists() then mission.ship:Undock()

into

    mission.ship:CallAt(mission.due, Ship.Undock)

Thoughts ?