Improve automatic star selection in the sector map
During manual movement in the sector map, when the option is enabled,
the closest star to the map viewing position is automatically selected.
But the search is performed only in the sector containing the original
position, so the wrong star may be highlighted. This commit fixes this
defect, now the search is performed in an increasing (if necessary)
range, and the really closest star is found.

Also, fixed that the star was not automatically selected when
animating "by inertia", but only when a key was pressed.