Re-implement Shields to use per-instance material
Instead of one global material which causes bugs with uninitialized memory, shields now have per-instance materials that have buffer data updated when needed.

Shield meshes by default are assigned to a static material that will not animate for cases when the model needs to be displayed without a Shields class instance attached.

Also adds a GetBufferBinding method to the UniformBuffer abstraction - this was missed due to the lack of externally-owned UniformBuffers during development.