Increase and limit the height of the BvhTree
The height of that tree can fluctuate very much, starting from log2(n)
and ending with n (n is the number of objects in space).

When traversing the tree, an array (for the stack) is used, the size of
which is fixed, because we do not want dynamic allocations in this case,
but must be at least the height of the tree.

We can limit the height of the tree, since a leaf can contain multiple objects.

It seems almost never in the game the height of a tree does not exceed 20.

So I took the number 38 (with a margin), it seems almost never the
height of the tree will be limited, and at the same time we are insured
against UB in all cases.