Fix PutShipOnRoute function
After placing the ship on the route, the possibility of a collision is
checked.  In case of a possible collision, its position is changed to
avoid this. It turns out that this calculation worked incorrectly,
because the FromVectors function creates a matrix using vectors as
columns, and in the further calculation it is assumed that vectors are
rows.

Also removed unnecessary Frame object retrieval by FrameId