The core of the economy.
'rn' is an implicit function of price and stock that determines ALL,
including the average profit from a ship that, in turn, determines
its sell value.

Current line determined profit from ships as constant.
This fact, in previous decade, effectively destroyed, and will undoubtedly
destroy in future, every attempt to fix trading from the fact that
it seems flat and boring.

The addition of an exponential not equal to 1, transforms the implicit
function of price and stock to invertible functions of one to the other.
Thus, this line becomes the ultimate mathematical tool by which the
developers can firmly control the economy consequences of their creations,
if they wish to.