Each ship appears in sell ship buletin board only if one of his
prerequisite ships is the player himself or if its prerequisite table is
empty.
Multiple chains of similar ships can be created by developers. Available
for buying will be the ships without prerequisites, the base and top
(without prerequisite) ship of each chain, along with player's ship and his
next in current chain ship. Thus, every chain of ships can be considered
his avatar. He can change chain but he is obliged to start from new base
to top all over again.

I can not create ships, so i used current ships as a chain, altering a
little their cargo and prices, in correspondence to the calibration i chosed
as best and close to current state of the game.

The present callibration can be played and it is strategically easier
without the prerequisite feature. From personal playing experience,
I highly recomment to play the callibrated economy along with
the prerequisite feature.

The correct place for the prerequisite tables of ships is in the
corresponding ship's json files. That would make lua code cleaner, simpler
and shorter. I am unable to even read C code, so i putted the prerequisite
tables in Calibration.lua file.