Dynamically change of reward according to ship's trading profit or to its number of missiles. Whoever hires player does not expect to pay the same reward, as he does when he hires smaller fighters, when he asks, large ships or the best of fighters, to get out of their usual schedule for his convenience.
In current state equipment value should be set according to their
possibilities for an increased profit. In that sense, as atmospheric
shields, radar, ECM etc does not participate in earning more profit but
instead, they weight some tones, decreasing trading profit from player,
player should be paid, and not pay, to equip them. So, the rewards
should be raised for better equipped ships.

That possibility, for the otherwise useless equipments as the above, is
not involved in calibration. It can be done in the future by a more
complete Calibration.Fighting function in Calibration.lua file.
(Or in Calibration.profit function, if it concerns trading ships).