With callibration of prices, the distribution in log space does not work as expected and the demand was always too low to be attractive.
The average reward of player is 27.5% of his investment per su. Mission
offers 100% in 2su. So, 50% per su per credit. Demand being at average
the half of Major export's average trading stock, determines the
average profit from that module to 25%. So, it comes that, if player
has sufficient for that commodity space, half of times module will
benefit the player and half of times it will not.