More transparent comm drop down window.
As player, most of the time i have system overview window open.
In current state, even if window is transparent with opacity=0.75,
i have the annoying feeling that something i miss from background.
With opacity=0.10 this feeling is much reduced.
In the rare case that background is so bright that i cannot see window's
informations, a hover of mouse upon it, as it uses other background,
reveals the missing information.