Fix PutShipOnRoute function (#5250)
* Fix PutShipOnRoute function

After placing the ship on the route, the possibility of a collision is
checked.  In case of a possible collision, its position is changed to
avoid this. It turns out that this calculation worked incorrectly,
because the FromVectors function creates a matrix using vectors as
columns, and in the further calculation it is assumed that vectors are
rows.

Also removed unnecessary Frame object retrieval by FrameId

The choice of trajectory correction method depends only on whether we
can calculate the tangent to the sphere of the effective radius of the
obstructor from the target. We cannot do this if the target is within
this sphere. All other conditions are unnecessary.