Change defaults.
Subjective options for a more earthly sky background. Easily tested and should be collaboratively decided.

AmountOfBackgroundStars 
Options do not allow 1.5 value i prefer. (Some times value of 1 becomes 0 on next startups of the game. I do not resolve this.) This parameter is sensitive for the results on screen. Previous 100 became 25 on last master but, because of other changes it seems again not enough reduction. I tested a lot on my screen and, for me, values  equal or larger than 2 fill background sky with too many stars for my mind to pick up constellations.

StarFieldStarSizeFactor
When, by previous AmountOfBackgroundStars, number of stars is reduced, big stars, that remain, appear clearly as unnaturally big and bright.