Remove duplicate function from propulsion
It turned out that AIModelCoordsMatchAngVel and AIMatchAngVelObjSpace
practically duplicate each other, with the exception of the softness
parameter, but this is a rather trivial difference. I left the function
written more concisely, in my opinion.

The removed function additionaly clamped the thrust value, which is
unnecessary, because it is clamped in the called SetAngThrusterState
anyway.