Add LuaObjectBase::AllocateNew
Add a templated function to allocate a new userdata sized for a specific C++ type and construct it in one line.
This reduces the amount of type-casting required to create an object and improves readability.

Added LuaPushPull overrides for string_view, size_t and uint64_t.
Use c++14 helper typename defs.