Fix getting exit parameters from hyperjump
The GetBodyForPath function was called (from GetHyperspaceExitParams)
before Game::m_state = State::NORMAL was set, so it returned nullptr.
Therefore, the first non-gravpoint body was always set as the base body
for exit.
This led to the fact that in the multiple system the star "A" was always
chosen, regardless of the player's choice in the sector map.