Add BodyComponent serialization
Components are serialized to JSON if present in a body object under their own namespace.

Components are automatically deserialized if registered; if the class loading constructor has created a component of the specified type already it will be used to deserialize JSON data into.
Otherwise, a new component of the specified type will be created and deserialized into.