Add Lua Components
LuaComponents are arbitrary table values that can be registered on Body objects to provide storage for module data that would otherwise have to be added in engine code.

LuaComponents currently do not have any kind of event handling and must be manually registered and deregistered from bodies by Lua code.