Add persistent LuaObject registry
C++-owned LuaObjects with externally-managed lifetimes don't need to be GC'd until the underlying object is destroyed.

LuaComponents depend on the handle for a given body remaining persistent as the handle controls the lifetime of the components.

Objects with lua-managed lifetimes (LuaSharedObject / LuaCopyObject / LuaOwnObject) remain in the transient registry and are GC'd when no references remain.