Ensure autosave before the onGameEnd event fires
Since the autosave at the end of the game was triggered by the onGameEnd
event, it was randomly queued among other subscribers to this event.
This resulted in (at least) an accidental loss of all TradeShips module
data during autosave when exiting the game.

Among other options, I chose to create a separate
'onAutoSaveBeforeGameEnds' event, as this seems to me the most
straightforward and clear.

Fixes #5427