Fix name generation in Search and Rescue
Addresses #6107

The name generatin in flavour 7 of the Search and Rescue module fails
when there is non-ASCII content or when there are more than two compo-
nents in the name. Fixed as follows. Sorting out names containing
non-ASCII characters

* Introduce a test for non-ASCII content at the start of the function
and returning the last component wich should be, if not 'the' surname,
at least 'a' surname.

The name is now first tested for 'suspicious' content and if it isn't
only clean printable ASCII, the string will be sorted on space, " ", as
delimiter and the last item returned. This has so far never failed to
do a correct split but will fail 100% with multiple surnames. The re-
sult will still be a proper name and as such useful. If the algorithm
doesn't find a space in the name either something is very wrong or the
character is a mononym. Pioneer doesn't have mononyms as far as I know.
In any case, should this happen, the full name is returned. (this may
look odd but a. not as odd as the worst cases today and b. it hasn't
failed so far and I've tested this a lot.

* Updates to the old algorithm

1 - Function now returns a string, not a table.

splitName() splits the name into it's components and stores them one by
one in a table which it returns on completed mission. It seems the
function was intended to have more general use to allow access to first
name and surname but this is hard to make in reality as names are quite
complex. A proper fix would mean a more fine grained access to the
names via the character. customer.name or target.surname. We can't
assume that the surname is number 2 in the table so splitName now
returns the surname and as a string.

2 - Count hyphen and apostrophe in the word

Some names use apostrophe in the middle of it. They are also not used
to split names between first name and surname so by doing this we make
both those groups statistically larger and the final split is pushed
more toward the middle of the name. Even if we get the final split
wrong the result will be more of a proper name.

3 - Return string two

If, after sorting into the list, there are only two elements, return
the second one. This will obviously always be correct.

4 - Use known indicators of surnames

Van Damme, de Lavall, etc. Instantly return anything after, and
including, this marker. This is a small list and also make a small
difference.

5 - Return the last element

Any remaining names are in the final station decided to have the final
element as it's surname. Slightly less than one name in a hundred reach
are undecided at this point and they all seam to be split correctly as
they are Latino double first names with one surname.