Use horizontal FOV in ModelSpinner
When setting the ModelSpinner projection, the FOV parameter means the
vertical viewing angle by default. This causes the zoom to be dependent
on the viewing height. Therefore, for example, in the ships market,
where the height of the widget with the ship is small, the ships are
drawn very small and there is a lot of free space on the sides.

Changed to make FOV stand for the horizontal field of view. Now, it
seems, the ships will be displayed more optimally.