Add new game start dialog
Visually, the dialog consists of widgets grouped into 4 tabs.  The
internal state is represented by a set of parameters sufficient to run
the game. Widgets do not exactly match parameters - there may be widgets
not represented by parameters, and parameters not shown as widgets, but
most parameters correspond to some widget.

Each parameter is an "instance" of some class, it can pull its data from
multiple sources (built-in start variant, savegame document), display an
interface for viewing/editing, validate it, and export the data to the
game's start routine.

Widgets have the property of blocking, by default editing most of the
parameters is not available. In order to unlock all parameters, there is
a 'Custom start' option, but it is disabled by default. When clicking on
the "New Game" button, if Ctrl is pressed, an additional option "Custom
start" will appear, unlocking all parameters.

In this commit, the dialog is only added, it is fully functional, but the
procedure for starting the game has not yet been implemented.