Start the game from a new dialog
Add one argument to the game constructor so that it immediately creates
a ship of the correct type, otherwise there were some problems with
geoms. If not docked, the ship is spawned in a low circular orbit.

Remove the ability to edit the face and name of the player inside the
game. Slightly reduced the image in personal information. Do not create
a player character at game start if one has already been created.