Let a player with autopilot off enter the orbital
Were previously added a separate type of collision meshes to the
orbital, flagged "ENTRANCE". These surfaces are identified by their name
beginning with "collision_port".

Now, upon collision with it, if player (or any other ship) have
autopilot turned on, docking occurs, if not, nothing happens. This is
necessary so that the player can, on the one hand, fly into the station
for manual docking, on the other hand, NPC ships and autopilot can use
the old docking mechanism.