Prevent zero length direction vector
Approximately once every one and a half in-game years (for the Sol
system), the following state may occur:

  - some ship is directed towards the planet
  - it has AICmdFlyTo command active
  - its target is on course
  - gravity, respectively, is directed along the the ship, and
    accelerates it
  - its reverse thrusters are _weaker than gravity_
  - it has zero speed

The combination of these circumstances leads to the fact that both
coefficients in the calculation of the head vector are equal to 0.
A vector of zero length is passed to Propulsion, where it is normalized
and SIGFPE occurs.