SystemView: introduce object track abstraction
Part one of a general decoupling of the map rendering from game logic. The map consumes a list of Projectables ("object tracks"), but is only responsible for generating and rendering SystemBody tracks for the current system being displayed.

Any external system (e.g. Sensors) can push tracks and orbits into the map to be displayed, including "fake" object tracks (e.g. sensor ghosts, e-war interference, etc.) using a unified interface.

Responsibility for understanding and rendering the contents of an unknown object track is reserved for the consumer of the final screenspace-positioned Projectables. This allows the map to be used in the context of the current game system, rendering cartography data on an unloaded system, or in an editor context.