Amend ship name generator list
Removed most "out of place" or trademarked names which readily associated with a fictional entity/character rather than a real name.
The focus is mostly on preserving the tone and context of the world of Pioneer rather than using 100% original names.
The criteria for remaining names is that they should not immediately stand out as belonging to a specific commercial work within the last ~50 years when combined with another name fragment.

Capitalized all name fragment lists, as the resulting generated word will be an english Proper Noun and should be capitalized.