Ignore x-axis only mouse wheel events
Some mice (like mine) have an x and y axis on the wheel.

Ignore mouse wheel events that are x-axis only (so therefore have a y-axis value of zero).  The behavior of my (Microsoft) mouse is that pushing left and right on the wheel repeatedly call mouse wheel y events as long as it's held.  We were interpreting these as always being mouse wheel down events, causing my clumsy fingers on the mouse wheel to zoom very quickly out in the third person ship view, sector view and so forth.