Make loading json arrays into lua more simple
This function has not been used yet, but will soon be useful for
restoring saves. But loading the array was strange. When loading an
array, it numbers indices starting from 0, and also adds the number of
array elements under the key 'n'.

Remove the additional field 'n' and make indexing from 1.