Add FlightLog as a start/recovery game parameter
We will simply show an additional tab in the start menu, no editing is
expected.

To format the string from the SystemPath, access to the galaxy generator
is required, this will be done using the Location tab, and also adding
the GetStarSystem method to LuaGalaxy.

If, for example, when restoring the flight log it turns out that such a
station no longer exists, we restore this as a system-only path or even
as a sector-only path. The flight log gui inside the game could not
handle this and crashed with an error, fixed.