Explicitly declare system bodies as orphans
When a StarSystem is deleted, we will reset all pointers to it in system
bodies, and this way we can determine that this system body is orphaned.
This will allow us to reliably distinguish such bodies and not
regenerate the system if it already exists.

Which, for example, in the SystemBody.parent attribute will allow us not
to access the global galaxy if we are sure that the StarSystem is not
deleted.