Update for pause menu
Changes:

1. Queuing of events onPauseMenuOpen/onPauseMenuClosed is moved to ui.optionsWindow:open() and ui.optionsWindow:close() respectively.
This was made to ensure queuing of this events no matter where this window was opened or closed;

2. Added ui.optionsWindow:changeState() method to group ui.optionsWindow:open() and ui.optionsWindow:close() into one;

3. Temporarily moved Input.EnableBindings(false) from onPauseMenuOpen, because when the game is paused, lua events are not processed.
Because of this, game will still read input bindings if pause menu was opened, while time acceleration was paused. This fixes it.

Bonus:
Save button will be blocked if player is hyperspacing.