Create a game settings tab in the settings view and add a difficulty slider to it
Difficulty is stored 2 ways, once for the Engine as a config value that will apply to any new game you start.

A game itself (so a savefile) will persist it's own difficulty and if it's not available on load, will use the engine configured value.

Also, separate out settings persistance from closing the window, so now if you open the settings window, change the settings and then save the game they are persisted in the saved game state.