Add updateParams fnc for tabs in the new game menu
If the 'updateLayout' function is called when setting/changing the size
of the window, so that all tabs can pre-calculate the required
dimensions of the elements, 'updateParams' is designed to update
dependent elements in accordance with (updated) GameParam.value.  The
function is guaranteed to run inside the ImGui frame, but we do not know
any dimensions. Neither one we want to run every frame, but the second
one might be more expensive and doesn't seem to need to be called if
we're just resizing. Although window resizing is not currently
supported, this may change in the future.