Fixup player.rating and player.reputation params
The numerical value of the rating turns out to be the number of kills,
so renamed player.rating to player.kills. And added a drag to
display/change the number of kills itself, interacting with the rating
combobox.

It turned out that before at startup these two parameters were not
passed to the game at all, fixed it.