Implement recover from saveload menu
If the saveload menu decides that it cannot load a save (most likely due
to a version mismatch), the Load button will be replaced with Recover
and you can try to restore the game.

In fact, this is the start of a new game, but we are trying to read the
parameters from the save. After a recovery attempt, the same dialog for
starting a new game opens, but without selecting a start option.

If an error occurs when reading a parameter from a save, or after
restoring the parameter is in an invalid state, it is unlocked so that
the player can manually specify it. The exception is the flight log - it
is always locked.

Version 89 is currently being successfully restored (2023-02-03). For
older versions, it is planned to add "reader" versions in the future.
Added the ability to create a "versioned" function; when it is launched,
dispatching occurs depending on the specified save version.

You can also force recovery by holding Ctrl in the boot menu. This
allows to restore a save with the same version as the current one. It
seems like this might be useful if changes are made between versions
that cause saves to be incompatible.

The input for recovery is not a raw document converted from json,
but a preprocessed one, namely all Lua objects are unpickled, and all
references are resolved.