Don't continue draw modals if current is closed
Inside the handler, this modal could be closed and removed along with
its children from the stack, while other modals could be created. And
the next modal on the stack may no longer be a child of this modal, but
a child of another modal that has not yet been opened.