Fix bug when grabbing screenshots that breaks menus with 3d renders embedded within
We were changing the buffer we read from to the front buffer, rather than the back buffer and not setting it back.

At the same time, modify that code to not try and do anything clever with stride as it's not needed and also to not grab the aplha channel as again, it's not needed.