Avoid rotating frame when starting in orbit
If you create a ship in a rotating frame corresponding to a body, in the
very next (time) frame it will be thrown into a non-rotating frame, and
the ship's speed will be changed, which will lead to a deviation of the
orbit from circular.

Also use body data for the orbital radius and not the system body, so as
not to accidentally fall under the terrain. For example, this sometimes
happened on Phobos.