Hardcode the project version
Since the build version is actually the date the build was run, if it is
rebuilt on a different day it becomes erroneous. For example, since in
any distribution the build into the image will most likely occur after
the release, or even rebuild it regularly, the version will differ from
the release one, which can be misleading.

There is also a second part of the version - the commit hash, but if the
stable version is (re)built from the source archive (which is correct),
there is no git repo at all.

So now it will be specified manually in CMakeLists, so that it
unambiguously indicates the version of the game.

Also made PROJECT_VERSION a cache variable, so that it is possible to
arbitrarily specify the version via command line arguments. It will be
useful for CI.

Also add PROJECT_VERSION_INFO cache variable. This variable is empty by
default. If it contains anything, it will be added to the commit hash in
the version, for example:

20240105 (340f22cae extra-info)

If the git repo is not available, only the string from the new variable
will be in parentheses.