WIP: Consolidate much flightlog ralated logic into the flightlog code and out of the new-game window.
The goal is to enact DRY, reducing the duplicated logic for rendering and serialzation of the flight log and cosolidating into one place.

So this:
 * Moved serialization (of which recovery is one version) of the flightlog into the flightlog code
 * Moves rendering of the flighog in the new-window screen to use the flight log rendering code

There are still some issues and ugly coupling remainng of the new-game window code and the flightog that needs unpicking around how we access the galaxy object to format log entries.  Perthaps we should do this once as the flight log entries are created, rather than on demand eavh frame of rendering.  This would help us clean up those dependencies.