Don't crash when there are few background stars
If the user sets a small AmountOfBackgroundStars, a small diameter ball
is scanned. So small that the number of tasks becomes comparable to the
height of this ball, and rounding problems begin to appear.

Rethought the intervals for tasks a little, now it seems to be fairly
uniform for such cases.

Also, if the user selects the number of stars to be 0, we simply do
nothing.