News Event: Fix commodity demand not being modified
When a news event causes a commodity to have 0 in stock,
and price multiplied, it should also boost demand. When stock
is in surplus, demand should be 0.

Demand linearly returns to "normal" as the hidden deadline for the event
approaches

Changes to market is applied when docking, and restored when un-docking