fix(savegame): fix savefiles directory not being created (#5917)
A refactor of the load/save UI in commit
7b6a79700682124d29351fb83562778eaf5a2da4 introduced a bug preventing the
"savefiles" directory from getting created, and causing an error-popup
whenever the directory was attempted to be accessed by the LUA UI.

The bug was resetting a variable to an unexpected value which then
caused the "sort" algorithm to throw an error. This was resolved by
ensuring that the variable always had a valid value for the sort
algorithm.