refactor: move and rearrange "edge buttons" in map views
Prepare for additional buttons to be added to the "edge buttons" area of
the map views.

Move the edge buttons to the top-right of the display, so all icons
remain in a fixed location irrespective of how many buttons there are.
This makes it easier to learn muscle-memory.

Also, rearrange the icons so that all 3 map views have the same list of
buttons from the top down, with any map-specific buttons following the
"Info" button.

Finally, remove the superfluous edge button to switch between System Map
and System Overview - they already exist on the top button bar.