feat: select between "current" and "selected" system view in maps
Add a button to switch between the Current System and the Selected System
in the System Map and the System Overview screens.

With this change, if the player has selected to view the
current system and is viewing one of the maps during a hyperspace jump,
a situation can occur in which SystemView::CalculateStarportHeight() is
called while the game state is in hyperspace. Hence a check is required
that the game is in normal space prior to calculating the real height of
a starport above a planet surface.