feat: ensure savegame directory is created on game startup
A fix was already made in commit b3282e30376c3ecce4624d2490bdc50d96559273
to ensure the savegame directory is created when saving a game. But when
starting the game for the first time and attempting to load a game, an
error message was shown that the savegame directory does not exist.

The savegame directory is now created on program startup. Also, all
savegame related logic in the C++ is moved into the "Game" class and out
of the "Pi" class. The savegame directory name is also fully encapsulated.
Ideally it would not be exposed at all, but some of the logging functions
use it to report on the full path of the savegame when errors occur.